+++ TombEngine/Objects/TR2/Vehicles/skidoo_info.h
@@ -7,9 +7,9 @@ namespace TEN::Entities::Vehicles
 		int LeftVerticalVelocity = 0;
 		int RightVerticalVelocity = 0;
 
-		short TurnRate = 0;
-		short MomentumAngle = 0;
-		short ExtraRotation = 0;
+		float TurnRate = 0;
+		float MomentumAngle = 0;
+		float ExtraRotation = 0;
 
 		int Pitch = 0;
 		int FlashTimer = 0;
